[iOS][web] Create interface for clearing browsing data
This CL only creates the backbone of the files and their interaction.
The idea is to move the clearing of the WK-related data from the chrome
layer to the web layer.
The actual implementation of the clear browsing data isn't done yet.

Bug: 619783
Change-Id: Ia396b5944afab00b74cae9d0e7f8c990a0c13685
Reviewed-on: https://chromium-review.googlesource.com/c/1470001
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#632577}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef4442a37e282fc61e9210b8590c02c31ca5099b