[iOS][Translate] Hides the translate infobar in fullscreen mode.
It does so by making the LegacyInfobarContainerViewController a
FullscreenUIElement and have it observe the FullscreenController. The VC
the passes the fullscreen progress to the TranslateInfobarView which is
also a FullscreenUIElement so it can hide/show itself accordingly.

Bug: 910994
Change-Id: I6fc7e28707820b077c7126c2fef9e6637403e3f3
Reviewed-on: https://chromium-review.googlesource.com/c/1471006
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#633132}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee506c5149cb648d1b2eaf47558c0ad4275c3c15